Extract tone handling to a new class
